Add time-shift test for renewer
[#113294537]
