+++ lib/utils/cluster/src/index.js
@@ -262,7 +262,7 @@ const worker = () => {
   // well and the master will terminate it
   const heartbeatReporter = setInterval(() => {
     process.send({
-      heartbeat: moment.valueOf()
+      heartbeat: moment.now()
     });
   }, heartbeatInterval);
