+++ lib/utils/moment/src/index.js
@@ -18,7 +18,7 @@ moment.now = () => {
 // much faster (factor of 10) due to native Date implementation
 //
 moment.toYYYYMM = (period) => {
-  const time = new Date(period * 86400000 + offset);
+  const time = new Date(period * 86400000);
   return time.getUTCFullYear() * 100 + time.getUTCMonth() + 1;
 };
