+++ lib/utils/moment/src/test/test.js
@@ -45,8 +45,8 @@ describe('abacus-moment', () => {
 
     it('converts days to YYYYMM format', () => {
       // 17223 days translates to the following date: 26-2-2017
-      // but since we have offset we expect the next month
-      expect(moment.toYYYYMM(17223)).to.equal(201703);
+      // we have offset, but we still expect the same month
+      expect(moment.toYYYYMM(17223)).to.equal(201702);
     });
   });
