+++ lib/utils/request/src/index.js
@@ -392,12 +392,15 @@ extend(request, {
 // corresponding responses
 let pings = [];
 
+const maxSuccessfulPings = 5;
+const counterBarrier = maxSuccessfulPings + 1;
+
 const ping = (uri, opt) => {
   const cb = function(err, val) {
     // Warning: mutating variable pings
     if(!err) {
       pings[t.uri].val = val;
-      pings[t.uri].count = (pings[t.uri].count || 0) + 1;
+      pings[t.uri].count = ((pings[t.uri].count || 0) + 1) % counterBarrier;
       return;
     }
 
@@ -429,8 +432,8 @@ const waitFor = (uri, opt, time, cb) => {
   const i = setInterval(() => {
     const n = request.ping(t.uri, t.options);
     if(n !== 0) debug('%d successful pings', n);
-    if(n === 5) {
-      // Call back after 5 successful pings
+    if(n === maxSuccessfulPings) {
+      // Call back after max successful pings
       clearInterval(i);
       clearTimeout(to);
       t.callback(undefined, t.uri);
