+++ lib/utils/request/src/test/test.js
@@ -475,80 +475,77 @@ describe('abacus-request', () => {
         });
       });
     });
-  
-  it('waits for request using specified timeout', (done) => {
+
+  context('when waiting on endpoint', () => {
     const timeout = 1000;
     const port = 3433;
 
-    // Create a test HTTP server
-    let gets = 0;
-    const server = http.createServer((req, res) => {
-      if(req.url === '/ok/request' && req.method === 'OPTIONS')
-      // Return an OK response with a body
-        res.end('okay');
-      else if(req.url === '/ok/request' && req.method === 'GET')
-      // Return an OK response with a body
-        res.end(gets++ === 0 ? 'okay' : 'notok');
-      else if(req.url === '/500/request' && req.method === 'POST') {
-        // Don't return a body here to test request's behavior
-        // without a body
-        res.statusCode = 500;
-        res.end();
-      }
-      else if(req.url === '/err/request' && req.method ===
-        'POST') {
-        // Return a JSON object with a message field, to test
-        // request's ability to pick up that message
-        let body = '';
-        req.on('data', (chunk) => {
-          body = body + chunk;
-        });
-        req.on('end', () => {
+    let server;
+
+    before(() => {
+      // Create a test HTTP server
+      let gets = 0;
+      server = http.createServer((req, res) => {
+        if (req.url === '/ok/request' && req.method === 'OPTIONS')
+        // Return an OK response with a body
+          res.end('okay');
+        else if (req.url === '/ok/request' && req.method === 'GET')
+        // Return an OK response with a body
+          res.end(gets++ === 0 ? 'okay' : 'notok');
+        else if (req.url === '/500/request' && req.method === 'POST') {
+          // Don't return a body here to test request's behavior
+          // without a body
           res.statusCode = 500;
-          res.setHeader('Content-type', 'application/json');
-          res.end(JSON.stringify({
-            message: JSON.parse(body)
-          }));
-        });
-      }
+          res.end();
+        }
+        else if (req.url === '/err/request' && req.method ===
+          'POST') {
+          // Return a JSON object with a message field, to test
+          // request's ability to pick up that message
+          let body = '';
+          req.on('data', (chunk) => {
+            body = body + chunk;
+          });
+          req.on('end', () => {
+            res.statusCode = 500;
+            res.setHeader('Content-type', 'application/json');
+            res.end(JSON.stringify({
+              message: JSON.parse(body)
+            }));
+          });
+        }
+      });
+
+      // Listen on an hard-coded port
+      setTimeout(() => server.listen(port), timeout);
     });
 
-    // Listen on an hard-coded port
-    setTimeout(() => server.listen(port), timeout);
-
-    let cbs = 0;
-    const verifyNumberOfRequests = () => {
-      if(++cbs === 5) done();
-    };
-
-    // Wait for timeout error
-    request.waitFor('http://localhost::p/ok/request', {
-      p: port
-    }, timeout / 2, (err, val) => {
-      expect(err).to.not.equal(undefined);
-      verifyNumberOfRequests();
+    after(() => {
+      server.close();
     });
 
-    // Wait for the server to become available
-    request.waitFor('http://localhost::p/ok/request', {
-      p: port
-    }, timeout * 5, (err, val) => {
-      if(err)
-        throw err;
+    it('uses the specified timeout', (done) => {
+      let cbs = 0;
+      const verifyNumberOfRequests = () => {
+        if (++cbs === 5) done();
+      };
 
-      // Send an HTTP request, expecting an OK response
-      request.get('http://localhost::p/:v/:r', {
-        cache: true,
-        p: port,
-        v: 'ok',
-        r: 'request'
-      }, (err, val) => {
-        expect(err).to.equal(undefined);
-        expect(val.statusCode).to.equal(200);
-        expect(val.body).to.equal('okay');
+      // Wait for timeout error
+      request.waitFor('http://localhost::p/ok/request', {
+        p: port
+      }, timeout / 2, (err, val) => {
+        expect(err).to.not.equal(undefined);
         verifyNumberOfRequests();
+      });
 
-        // Send an HTTP request, expecting a cached OK response
+      // Wait for the server to become available
+      request.waitFor('http://localhost::p/ok/request', {
+        p: port
+      }, timeout * 5, (err, val) => {
+        if (err)
+          throw err;
+
+        // Send an HTTP request, expecting an OK response
         request.get('http://localhost::p/:v/:r', {
           cache: true,
           p: port,
@@ -559,32 +556,56 @@ describe('abacus-request', () => {
           expect(val.statusCode).to.equal(200);
           expect(val.body).to.equal('okay');
           verifyNumberOfRequests();
+
+          // Send an HTTP request, expecting a cached OK response
+          request.get('http://localhost::p/:v/:r', {
+            cache: true,
+            p: port,
+            v: 'ok',
+            r: 'request'
+          }, (err, val) => {
+            expect(err).to.equal(undefined);
+            expect(val.statusCode).to.equal(200);
+            expect(val.body).to.equal('okay');
+            verifyNumberOfRequests();
+          });
+        });
+
+        // Send an HTTP request, expecting a 500 status code
+        // Here test the option to pass the URI as a field of the options object
+        request.post({
+          uri: 'http://localhost::p/:v/:r',
+          p: port,
+          v: '500',
+          r: 'request'
+        }, (err, val) => {
+          expect(err.message).to.equal('HTTP response status code 500');
+          expect(val).to.equal(undefined);
+          verifyNumberOfRequests();
+        });
+
+        // Send an HTTP request, expecting an error message
+        request.post('http://localhost::p/:v/:r', {
+          p: port,
+          v: 'err',
+          r: 'request',
+          body: 'duh'
+        }, (err, val) => {
+          expect(err.message).to.equal('duh');
+          expect(val).to.equal(undefined);
+          verifyNumberOfRequests();
         });
       });
+    });
 
-      // Send an HTTP request, expecting a 500 status code
-      // Here test the option to pass the URI as a field of the options object
-      request.post({
-        uri: 'http://localhost::p/:v/:r',
-        p: port,
-        v: '500',
-        r: 'request'
-      }, (err, val) => {
-        expect(err.message).to.equal('HTTP response status code 500');
-        expect(val).to.equal(undefined);
-        verifyNumberOfRequests();
-      });
+    it('resets ping counts', (done) => {
+      const url = 'http://localhost::p/ok/request';
 
-      // Send an HTTP request, expecting an error message
-      request.post('http://localhost::p/:v/:r', {
-        p: port,
-        v: 'err',
-        r: 'request',
-        body: 'duh'
-      }, (err, val) => {
-        expect(err.message).to.equal('duh');
-        expect(val).to.equal(undefined);
-        verifyNumberOfRequests();
+      // Wait should pass since ping count is reset to 0
+      request.waitFor(url, { p: port }, (err) => {
+        expect(err).to.equal(undefined);
+
+        done();
       });
     });
   });
